requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/base.in requirements/dev.in requirements/quality.in
#
argparse==1.4.0           # via caniusepython3
backports.functools-lru-cache==1.4  # via caniusepython3
caniusepython3==6.0.0
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
configparser==3.5.0       # via pydocstyle
diff-cover==1.0.2
distlib==0.2.6            # via caniusepython3
djangorestframework==3.6.3
django-model-utils==3.0.0
edx-i18n-tools==0.4.3
edx-lint==0.5.4
edx-opaque-keys==0.4.1
first==2.0.1              # via pip-tools
futures==3.2.0            # via caniusepython3
idna==2.6                 # via requests
inflect==0.2.5            # via jinja2-pluralize
isort==4.2.15
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10              # via diff-cover, jinja2-pluralize
markupsafe==1.0           # via jinja2
packaging==16.8           # via caniusepython3
path.py==10.5             # via edx-i18n-tools
pip-tools==1.11.0
pkginfo==1.4.1            # via twine
pluggy==0.6.0             # via tox
polib==1.1.0              # via edx-i18n-tools
py==1.5.2                 # via tox
pycodestyle==2.3.1
pydocstyle==2.1.1
pygments==2.2.0           # via diff-cover
pyparsing==2.2.0          # via packaging
pytz==2016.7
pyyaml==3.12              # via edx-i18n-tools
requests-toolbelt==0.8.0  # via twine
requests==2.18.4          # via caniusepython3, requests-toolbelt, twine
six==1.11.0               # via diff-cover, edx-i18n-tools, packaging, pip-tools, pydocstyle, tox
snowballstemmer==1.2.1    # via pydocstyle
tox-battery==0.2
tox==2.9.1
tqdm==4.19.5              # via twine
twine==1.9.1
urllib3==1.22             # via requests
virtualenv==15.1.0        # via tox
wheel==0.30.0
XBlock==1.1.1
